fix(objc): keep playground samples in preview source
